+++ src/commands/local/ps.ts
@@ -15,7 +15,9 @@
  */
 
 import { BaseCommand, baseFlags } from "../../baseCommand.js";
-import { initNewDockerComposeConfig } from "../../lib/configs/project/dockerCompose.js";
+import { commandObj } from "../../lib/commandObj.js";
+import { initReadonlyDockerComposeConfig } from "../../lib/configs/project/dockerCompose.js";
+import { DOCKER_COMPOSE_FULL_FILE_NAME } from "../../lib/const.js";
 import { dockerCompose } from "../../lib/dockerCompose.js";
 import { initCli } from "../../lib/lifeCycle.js";
 
@@ -27,7 +29,13 @@ export default class PS extends BaseCommand<typeof PS> {
   };
   async run(): Promise<void> {
     await initCli(this, await this.parse(PS));
-    const dockerComposeConfig = await initNewDockerComposeConfig();
+    const dockerComposeConfig = await initReadonlyDockerComposeConfig();
+
+    if (dockerComposeConfig === null) {
+      commandObj.error(
+        `Cannot find ${DOCKER_COMPOSE_FULL_FILE_NAME}. Aborting.`,
+      );
+    }
 
     await dockerCompose({
       args: ["ps"],