+++ src/lib/configs/project/dockerCompose.ts
@@ -15,13 +15,13 @@
  */
 
 import assert from "assert";
+import { join, relative } from "path";
 
 import type { JSONSchemaType } from "ajv";
 import { yamlDiffPatch } from "yaml-diff-patch";
 
 import { versions } from "../../../versions.js";
 import {
-  LOCAL_IPFS_ADDRESS,
   DOCKER_COMPOSE_FILE_NAME,
   DOCKER_COMPOSE_FULL_FILE_NAME,
   TOP_LEVEL_SCHEMA_ID,
@@ -32,10 +32,11 @@ import {
   TCP_PORT_START,
   WEB_SOCKET_PORT_START,
   PROVIDER_CONFIG_FULL_FILE_NAME,
+  CONFIGS_DIR_NAME,
 } from "../../const.js";
 import type { ProviderConfigArgs } from "../../generateUserProviderConfig.js";
 import { getSecretKeyOrReturnExisting } from "../../keyPairs.js";
-import { getFluenceDir } from "../../paths.js";
+import { ensureFluenceConfigsDir, getFluenceDir } from "../../paths.js";
 import {
   getConfigInitFunction,
   getReadonlyConfigInitFunction,
@@ -47,8 +48,10 @@ import {
 
 import {
   initNewReadonlyProviderConfig,
+  ensureConfigToml,
   type ProviderConfigReadonly,
 } from "./provider.js";
+import { getConfigTomlName } from "./provider.js";
 
 type Service = {
   image?: string;
@@ -144,8 +147,6 @@ const configSchemaV0: JSONSchemaType<ConfigV0> = {
   required: ["version", "services"],
 };
 
-const NOX_IPFS_MULTIADDR = `/dns4/${IPFS_CONTAINER_NAME}/tcp/${IPFS_PORT}`;
-
 type GenNoxImageArgs = {
   name: string;
   tcpPort: number;
@@ -161,6 +162,8 @@ function genNox({
   bootstrapName,
   bootstrapTcpPort,
 }: GenNoxImageArgs): [name: string, service: Service] {
+  const configTomlName = getConfigTomlName(name);
+  const configLocation = `/run/${CONFIGS_DIR_NAME}/${configTomlName}`;
   return [
     name,
     {
@@ -168,28 +171,14 @@ function genNox({
       pull_policy: "always",
       ports: [`${tcpPort}:${tcpPort}`, `${webSocketPort}:${webSocketPort}`],
       environment: {
-        FLUENCE_ENV_AQUA_IPFS_EXTERNAL_API_MULTIADDR: LOCAL_IPFS_ADDRESS,
-        FLUENCE_ENV_DECIDER_IPFS_MULTIADDR: NOX_IPFS_MULTIADDR,
-        FLUENCE_ENV_AQUA_IPFS_LOCAL_API_MULTIADDR: NOX_IPFS_MULTIADDR,
-        FLUENCE_ENV_CONNECTOR_API_ENDPOINT: `http://${CHAIN_CONTAINER_NAME}:${CHAIN_PORT}`,
-        FLUENCE_ENV_CONNECTOR_FROM_BLOCK: "earliest",
         WASM_LOG: "info",
         RUST_LOG:
           "debug,particle_reap=debug,aquamarine=warn,aquamarine::particle_functions=debug,aquamarine::log=debug,aquamarine::aqua_runtime=error,ipfs_effector=off,ipfs_pure=off,system_services=debug,marine_core::module::marine_module=info,tokio_threadpool=info,tokio_reactor=info,mio=info,tokio_io=info,soketto=info,yamux=info,multistream_select=info,libp2p_secio=info,libp2p_websocket::framed=info,libp2p_ping=info,libp2p_core::upgrade::apply=info,libp2p_kad::kbucket=info,cranelift_codegen=info,wasmer_wasi=info,cranelift_codegen=info,wasmer_wasi=info,run-console=trace,wasmtime_cranelift=off,wasmtime_jit=off,libp2p_tcp=off,libp2p_swarm=off,particle_protocol::libp2p_protocol::upgrade=info,libp2p_mplex=off,particle_reap=off,netlink_proto=warn",
-        FLUENCE_SYSTEM_SERVICES__ENABLE: "aqua-ipfs,decider",
-        FLUENCE_ENV_CONNECTOR_WALLET_KEY:
-          "0xfdc4ba94809c7930fe4676b7d845cbf8fa5c1beae8744d959530e5073004cf3f",
-        FLUENCE_ENV_CONNECTOR_CONTRACT_ADDRESS:
-          "0x0f68c702dC151D07038fA40ab3Ed1f9b8BAC2981",
-        FLUENCE_SYSTEM_SERVICES__DECIDER__DECIDER_PERIOD_SEC: 10,
         FLUENCE_MAX_SPELL_PARTICLE_TTL: "9s",
-        FLUENCE_SYSTEM_SERVICES__DECIDER__NETWORK_ID: 31337,
         FLUENCE_ROOT_KEY_PAIR__PATH: `/run/secrets/${name}`,
       },
       command: [
-        "--aqua-pool-size=5",
-        `-t=${tcpPort}`,
-        `-w=${webSocketPort}`,
+        `--config=${configLocation}`,
         "--external-maddrs",
         `/dns4/${name}/tcp/${tcpPort}`,
         `/dns4/${name}/tcp/${webSocketPort}/ws`,
@@ -199,6 +188,7 @@ function genNox({
           : `--bootstraps=/dns/${bootstrapName}/tcp/${bootstrapTcpPort}`,
       ],
       depends_on: [IPFS_CONTAINER_NAME],
+      volumes: [`./${CONFIGS_DIR_NAME}/${configTomlName}:${configLocation}`],
       secrets: [name],
     },
   ];
@@ -207,16 +197,23 @@ function genNox({
 async function genDockerCompose(
   providerConfig: ProviderConfigReadonly,
 ): Promise<LatestConfig> {
+  const configsDir = await ensureFluenceConfigsDir();
+  const fluenceDir = getFluenceDir();
+
   const peers = await Promise.all(
-    Object.entries(providerConfig.computePeers).map(
-      async ([name, { webSocketPort, tcpPort }]) => {
-        return {
-          ...(await getSecretKeyOrReturnExisting(name)),
-          webSocketPort,
-          tcpPort,
-        };
-      },
-    ),
+    Object.entries(providerConfig.computePeers).map(async ([name, { nox }]) => {
+      const relativeConfigFilePath = relative(
+        fluenceDir,
+        join(configsDir, getConfigTomlName(name)),
+      );
+
+      return {
+        ...(await getSecretKeyOrReturnExisting(name)),
+        webSocketPort: nox?.websocketPort,
+        tcpPort: nox?.tcpPort,
+        relativeConfigFilePath,
+      };
+    }),
   );
 
   const [bootstrap, ...restNoxes] = peers;
@@ -237,12 +234,10 @@ async function genDockerCompose(
     services: {
       [CHAIN_CONTAINER_NAME]: {
         image: versions.chain,
-        pull_policy: "always",
         ports: [`${CHAIN_PORT}:${CHAIN_PORT}`],
       },
       [IPFS_CONTAINER_NAME]: {
         image: "ipfs/go-ipfs",
-        pull_policy: "always",
         ports: [`${IPFS_PORT}:${IPFS_PORT}`, "4001:4001"],
         environment: {
           IPFS_PROFILE: "server",
@@ -270,8 +265,8 @@ async function genDockerCompose(
       ),
     },
     secrets: Object.fromEntries(
-      peers.map(({ name, path: file }) => {
-        return [name, { file }];
+      peers.map(({ name, relativeSecretFilePath: file }) => {
+        return [name, { file }] as const;
       }),
     ),
   };
@@ -303,10 +298,9 @@ const initConfigOptions = {
   getConfigOrConfigDirPath: getFluenceDir,
 };
 
-export async function initNewDockerComposeConfig(
-  args: ProviderConfigArgs = {},
-) {
+export async function initNewDockerComposeConfig(args: ProviderConfigArgs) {
   const providerConfig = await initNewReadonlyProviderConfig(args);
+  await ensureConfigToml(providerConfig);
   return getConfigInitFunction(
     initConfigOptions,
     await genDefaultDockerCompose(providerConfig),
@@ -314,9 +308,10 @@ export async function initNewDockerComposeConfig(
 }
 
 export async function initNewReadonlyDockerComposeConfig(
-  args: ProviderConfigArgs = {},
+  args: ProviderConfigArgs,
 ) {
   const providerConfig = await initNewReadonlyProviderConfig(args);
+  await ensureConfigToml(providerConfig);
   return getReadonlyConfigInitFunction(
     initConfigOptions,
     await genDefaultDockerCompose(providerConfig),